+++ test/src/compare_pose.cpp
@@ -58,10 +58,10 @@ TEST(ComparePose, Compare)
       const float x_error = path.poses[i_path].pose.position.x - msg->pose.pose.position.x;
       const float y_error = path.poses[i_path].pose.position.y - msg->pose.pose.position.y;
       const float z_error = path.poses[i_path].pose.position.z - msg->pose.pose.position.z;
-      const float error = sqrtf(powf(x_error, 2.0) + powf(y_error, 2.0) + powf(z_error, 2.0));
-      const float x_sigma = sqrtf(msg->pose.covariance[0 * 6 + 0]);
-      const float y_sigma = sqrtf(msg->pose.covariance[1 * 6 + 1]);
-      const float z_sigma = sqrtf(msg->pose.covariance[2 * 6 + 2]);
+      const float error = std::sqrt(std::pow(x_error, 2.0) + std::pow(y_error, 2.0) + std::pow(z_error, 2.0));
+      const float x_sigma = std::sqrt(msg->pose.covariance[0 * 6 + 0]);
+      const float y_sigma = std::sqrt(msg->pose.covariance[1 * 6 + 1]);
+      const float z_sigma = std::sqrt(msg->pose.covariance[2 * 6 + 2]);
 
       fprintf(stderr, "compare_pose[%lu/%lu]:\n",
               i_path, path.poses.size());
